Possibly fix linux warning

+++ PlayMode.cpp
@@ -477,7 +477,7 @@ void PlayMode::draw(glm::uvec2 const &drawable_size) {
             }
 
             //skip remaining lights if maximum light count reached:
-            if (light_type.size() == lights) break;
+            if (light_type.size() == (uint32_t)lights) break;
         }
 
         GL_ERRORS();
